refactor(inference): extract distributed semaphore to services/inference (phase 6D)
Move DistributedSemaphoreActor and DistributedSemaphore from
components/utils.py to services/inference/distributed_semaphore.py.
Remove module-level config read from @ray.remote decorator.
Old locations re-export from the new canonical home.
