+++ openrag/services/inference/distributed_semaphore.py
@@ -1,0 +1,61 @@
+"""Ray-based distributed semaphore for cluster-wide concurrency limiting.
+
+Extracted from ``components/utils.py``.  The actor handles acquire/release;
+``DistributedSemaphore`` locates (or creates) the actor and wraps it as an
+async context manager.
+"""
+
+from __future__ import annotations
+
+import asyncio
+
+import ray
+
+
+@ray.remote(max_restarts=5)
+class DistributedSemaphoreActor:
+    def __init__(self, max_concurrent_ops: int):
+        self.semaphore = asyncio.Semaphore(max_concurrent_ops)
+
+    async def acquire(self):
+        await self.semaphore.acquire()
+
+    def release(self):
+        self.semaphore.release()
+
+
+class DistributedSemaphore:
+    """Async context manager backed by a detached Ray actor.
+
+    The actor is created on first use (get-or-create) and survives across
+    callers within the same Ray cluster.
+    """
+
+    def __init__(
+        self,
+        name: str = "llmSemaphore",
+        namespace: str = "openrag",
+        max_concurrent_ops: int = 10,
+    ):
+        self._name = name
+        self._namespace = namespace
+        self._max_concurrent_ops = max_concurrent_ops
+
+    def _get_or_create_actor(self):
+        try:
+            return ray.get_actor(self._name, namespace=self._namespace)
+        except ValueError:
+            return DistributedSemaphoreActor.options(
+                name=self._name,
+                namespace=self._namespace,
+                lifetime="detached",
+            ).remote(self._max_concurrent_ops)
+
+    async def __aenter__(self):
+        semaphore_actor = self._get_or_create_actor()
+        await semaphore_actor.acquire.remote()
+        return self
+
+    async def __aexit__(self, exc_type, exc, tb):
+        semaphore_actor = self._get_or_create_actor()
+        await semaphore_actor.release.remote()
